re-organize main classes
